telegram-bot/requirements.txt
requests
python-telegram-bot
python-decouple
